refactor(server): modularize routes
